backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
celery==5.3.6
pydantic==2.12.4
pydantic-settings==2.12.0
python-multipart==0.0.9
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
authlib==1.3.0
itsdangerous==2.1.2
chromadb==0.5.5
openai==2.8.1
httpx==0.28.1
unstructured[pdf]==0.18.20
pdf2image==1.17.0
tiktoken==0.7.0
google-genai==1.52.0
beautifulsoup4==4.12.3
pandas==2.2.3
tabulate==0.9.0
ragas==0.3.9
datasets==4.4.1
langchain-google-genai==3.2.0
langchain-openai==1.1.0
pytest==8.3.3
redis==7.1.0
prometheus-client==0.20.0
python-json-logger==2.0.7
PyYAML==6.0.2
sentry-sdk==1.40.5
email-validator==2.3.0
greenlet==3.0.3
jieba==0.42.1
rank_bm25